@@ -1,12 +1,13 @@
 package com.news.NS.controller;
 
+import cn.dev33.satoken.stp.StpUtil;
 import cn.dev33.satoken.annotation.SaCheckRole;
 import cn.dev33.satoken.annotation.SaMode;
-import cn.dev33.satoken.stp.StpUtil;
+import com.news.NS.common.CommonConstant;
 import com.news.NS.common.ResponseBodyResult;
 import com.news.NS.common.domain.PageInfo;
 import com.news.NS.domain.User;
-import com.news.NS.domain.dto.*;
+import com.news.NS.domain.dto.User.*;
 import com.news.NS.service.UserService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -86,6 +87,7 @@ public class UserController {
 
     @PostMapping("/admin/update/role")
     @ApiOperation(value = "更改用户权限")
+    @SaCheckRole(value = {CommonConstant.ADMIN,CommonConstant.SUPER_ADMIN},mode = SaMode.OR)
     public void changeRoleApi(@RequestBody UserRoleChangeDTO userRoleChangeDTO) {
         userService.changeUserRole(userRoleChangeDTO.getUserId(), userRoleChangeDTO.getIdentification());
     }
