@@ -1,12 +1,13 @@
-package com.news.NS.domain.dto;
+package com.news.NS.domain.dto.Comment;
 
 import lombok.Data;
 import org.hibernate.validator.constraints.Range;
 
 import javax.validation.constraints.Min;
+import javax.validation.constraints.NotNull;
 
 @Data
-public class CommentListAdminQueryDTO {
+public class CommentListQueryDTO {
 
     @Min(1)
     private Integer page;
@@ -14,5 +15,6 @@ public class CommentListAdminQueryDTO {
     @Range(min = 1, max = 100)
     private Integer size;
 
+    @NotNull
     private Integer newsId;
 }
