@@ -1,7 +1,8 @@
-package com.news.NS.domain.dto;
+package com.news.NS.domain.dto.User;
 
 import lombok.Data;
 import lombok.ToString;
+import org.hibernate.validator.constraints.Length;
 
 /**
  * @author 车
@@ -11,8 +12,10 @@ import lombok.ToString;
 @Data
 @ToString
 public class UserUpdatePwdDTO {
+    @Length(min = 11,max = 11)
     private String tel;
+    @Length(min = 7, max = 20)
     private String newPwd;
-
+    @Length(min = 6,max = 6)
     private String code;
 }
